requirements.txt
# Financial Document Intelligence Platform - Requirements
# Core ML
torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
sentence-transformers>=2.2.2

# Vector Store
qdrant-client>=1.7.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pdfplumber>=0.10.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
tabulate>=0.9.0


# Search
rank-bm25>=0.2.2

# API & UI
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.0
PyYAML>=6.0.1

# Evaluation
datasets>=2.15.0
evaluate>=0.4.1
rouge-score>=0.1.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
